pyproject.toml
[project]
name = "open-mirroring-debezium"
version = "0.1.0"
description = "Debezium CDC events from EventHub → Parquet → Microsoft Fabric Open Mirroring landing zone"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Luke Moloney" }]

dependencies = [
    "azure-eventhub>=5.11",
    "azure-eventhub-checkpointstoreblob-aio>=1.1",
    "azure-storage-file-datalake>=12.14",
    "azure-identity>=1.15",
    "pyarrow>=15.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "pyright>=1.1",
]

[project.scripts]
omd-consumer = "open_mirroring_debezium.consumer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_mirroring_debezium"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
